CodemyQirk/opApi/requirements.txt
apturl==0.5.2
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
Brlapi==0.8.3
certifi==2025.6.15
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.2
click==8.0.3
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
constantly==23.10.4
cryptography==45.0.5
cupshelpers==1.0
daphne==4.2.1
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.2
duplicity==0.8.21
fasteners==0.14.1
filelock==3.6.0
future==0.18.2
gast==0.5.2
git-filter-repo==2.38.0
gpg==1.16.0
html5lib==1.1
httpie==3.2.4
httplib2==0.20.2
humanfriendly==10.0
hyperlink==21.0.0
idna==3.10
imageio==2.31.1
img2pdf==0.4.4
importlib-metadata==4.6.4
imutils==0.5.4
incremental==24.7.2
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
libcomps==0.1.15
lockfile==0.12.2
louis==3.20.0
lxml==4.9.3
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
multidict==6.6.3
netifaces==0.11.0
networkx==3.1
notify2==0.3
numpy==1.24.3
oauthlib==3.2.0
ocrmypdf==14.3.0
olefile==0.46
opencv-python==4.8.0.76
openshot-qt==2.5.1
packaging==23.1
PAM==0.4.2
paramiko==2.9.3
pdfminer.six==20221105
pexpect==4.8.0
pikepdf==8.2.3
Pillow==9.0.1
platformdirs==2.5.1
pluggy==1.2.0
ply==3.11
protobuf==3.12.4
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
Pygments==2.19.2
PyGObject==3.42.1
PyICU==2.8.1
PyJWT==2.9.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==25.1.0
pyparsing==2.4.7
PyPDF2==3.0.1
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
pythran==0.10.0
pytz==2022.1
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rpm==4.17.0
scikit-image==0.21.0
scipy==1.10.1
scour==0.38.2
SecretStorage==3.3.1
service-identity==24.2.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.5.3
ssh-import-id==5.11
systemd-python==234
tesserocr==2.6.1
tifffile==2023.7.18
tomli==2.2.1
tqdm==4.66.0
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.5.0
usb-creator==0.3.7
virtualenv==20.13.0+ds
wadllib==1.3.6
webencodings==0.5.1
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==7.2
